moved mock into test directory
