+++ package.json
@@ -5,7 +5,7 @@
   "main": "index.js",
   "scripts": {
     "start": "node index",
-    "test": "NODE_ENV=test istanbul cover --report html node_modules/.bin/_mocha --recursive -- -R spec tests/**/*.js"
+    "test": "NODE_ENV=test istanbul cover --report html node_modules/.bin/_mocha --recursive -- -R spec tests/*.js"
   },
   "repository": {
     "type": "git",
