+++ tests/mock/imgurGallery.js
