+++ tests/mock/imgurSearch.js
