+++ tests/mock/imgurServiceErrors.js
